moved constant into config
