+++ packages/extension-chakra-storefront/config/sample.json
@@ -3,7 +3,6 @@
     "activeDataEnabled": false,
     "maxCacheAge": 900,
     "staleWhileRevalidate": 900,
-    "defaultSiteTitle": "Retail React App",
     "commerceAPI": {
         "proxyPath": "/mobify/proxy/api",
         "parameters": {
@@ -15,6 +14,7 @@
     },
     "defaultSite": "RefArchGlobal",
     "defaultLocale": "en-US",
+    "defaultSiteTitle": "Retail React App",
     "einsteinAPI": {
         "host": "https://api.cquotient.com",
         "einsteinId": "1ea06c6e-c936-4324-bcf0-fada93f83bb1",
