+++ packages/extension-chakra-storefront/src/components/with-layout/with-layout.tsx
@@ -40,21 +40,14 @@ import ScrollToTop from '../scroll-to-top'
 // Local Project Hooks
 import {AuthModal, useAuthModal} from '../../hooks/use-auth-modal'
 import {AddToCartModalProvider} from '../../hooks/use-add-to-cart-modal'
-import {useExtensionConfig} from '../../hooks/use-extension-config'
-import {useCurrentCustomer} from '../../hooks/use-current-customer'
-import {useCurrentBasket} from '../../hooks/use-current-basket'
+import {useExtensionConfig, useCurrentCustomer, useCurrentBasket} from '../../hooks'
 import {watchOnlineStatus, flatten} from '../../utils/utils'
 import useActiveData from '../../hooks/use-active-data'
 import useMultiSite from '../../hooks/use-multi-site'
+import {useTheme} from '@chakra-ui/react'
 
 // CONSTANTS
-import {
-    DEFAULT_SITE_TITLE,
-    HOME_HREF,
-    THEME_COLOR,
-    CAT_MENU_DEFAULT_NAV_SSR_DEPTH,
-    CAT_MENU_DEFAULT_ROOT_CATEGORY
-} from '../../constants'
+import {CAT_MENU_DEFAULT_NAV_SSR_DEPTH, CAT_MENU_DEFAULT_ROOT_CATEGORY} from '../../constants'
 
 // Define a type for the HOC props
 type WithAppLayoutProps = React.ComponentPropsWithoutRef<any>
@@ -116,7 +109,7 @@ const withLayout = <P extends object>(WrappedComponent: React.ComponentType<P>) 
         const {site, locale, buildUrl} = useMultiSite()
         const [isOnline, setIsOnline] = useState(true)
         const styles = useStyleConfig('App')
-
+        const {colors} = useTheme()
         const {isOpen, onOpen, onClose} = useDisclosure()
 
         // Used to conditionally render header/footer for checkout page
@@ -176,7 +169,7 @@ const withLayout = <P extends object>(WrappedComponent: React.ComponentType<P>) 
 
         const onLogoClick = () => {
             // Goto the home page.
-            const path = buildUrl(HOME_HREF)
+            const path = buildUrl(config?.pages.Home.path)
 
             history.push(path)
 
@@ -227,7 +220,7 @@ const withLayout = <P extends object>(WrappedComponent: React.ComponentType<P>) 
                 </Helmet>
 
                 <Seo>
-                    <meta name="theme-color" content={THEME_COLOR} />
+                    <meta name="theme-color" content={colors.blue['600']} />
                     <meta name="apple-mobile-web-app-title" content={config.defaultSiteTitle} />
 
                     {/* Urls for all localized versions of this page (including current page)
