+++ packages/extension-chakra-storefront/src/components/with-react-intl/with-react-intl.tsx
@@ -10,15 +10,12 @@ import {IntlProvider} from 'react-intl'
 import {useQuery} from '@tanstack/react-query'
 import {useLocation} from 'react-router-dom'
 
-// CONSTANTS
-import {DEFAULT_LOCALE} from '../../constants'
-
 // Local Imports
 import {getTargetLocale, fetchTranslations} from '../../utils/locale'
 import {isServer} from '../../utils/utils'
 import logger from '../../utils/logger-instance'
 import useMultiSite from '../../hooks/use-multi-site'
-import {useExtensionConfig} from '../..//hooks/use-extension-config'
+import {useExtensionConfig} from '../../hooks'
 
 // Define a type for the HOC props
 type WithReactIntlProps = React.ComponentPropsWithoutRef<any>
@@ -42,7 +39,7 @@ const withReactIntl = <P extends object>(WrappedComponent: React.ComponentType<P
                 // then the app would use the default locale as the fallback.
 
                 // NOTE: Your implementation may differ, this is just what we did.
-                return [locale?.id || config.DEFAULT_LOCALE]
+                return [locale?.id || config.defautLocale]
             },
             l10nConfig: site.l10n
         })
@@ -93,7 +90,7 @@ const withReactIntl = <P extends object>(WrappedComponent: React.ComponentType<P
                 // NOTE: if you update this value, please also update the following npm scripts in `template-retail-react-app/package.json`:
                 // - "extract-default-translations"
                 // - "compile-translations:pseudo"
-                defaultLocale={DEFAULT_LOCALE}
+                defaultLocale={config.defaultLocale}
             >
                 <WrappedComponent {...(props as P)} />
             </IntlProvider>
