+++ packages/extension-chakra-storefront/src/hooks/index.js
@@ -16,3 +16,5 @@ export {useVariationParams} from './use-variation-params'
 export {useDerivedProduct} from './use-derived-product'
 export {useCurrency} from './use-currency'
 export {useExtensionConfig} from './use-extension-config'
+export {useCurrentCustomer} from './use-current-customer'
+export {useCurrentBasket} from './use-current-basket'
