+++ packages/template-typescript-minimal/package.json
@@ -1,81 +1,82 @@
 {
-  "name": "typescript-minimal",
-  "version": "4.0.0-extensibility-preview.3",
-  "private": true,
-  "scripts": {
-    "analyze-build": "cross-env MOBIFY_ANALYZE=true npm run build",
-    "build": "pwa-kit-dev build",
-    "format": "pwa-kit-dev format \"**/*.{js,jsx,ts,tsx}\"",
-    "lint": "npm run lint:js && npm run typecheck",
-    "lint:fix": "npm run lint:js -- --fix",
-    "lint:js": "pwa-kit-dev lint \"**/*.{js,jsx,ts,tsx}\"",
-    "push": "npm run build && pwa-kit-dev push",
-    "save-credentials": "pwa-kit-dev save-credentials",
-    "start": "pwa-kit-dev start",
-    "start:inspect": "npm run start -- --inspect",
-    "tail-logs": "pwa-kit-dev tail-logs",
-    "test": "pwa-kit-dev test",
-    "typecheck": "tsc --noEmit"
-  },
-  "devDependencies": {
-    "@chakra-ui/react": "^2.8.2",
-    "@emotion/react": "^11.13.3",
-    "@emotion/styled": "^11.13.0",
-    "@loadable/component": "^5.15.3",
-    "@salesforce/pwa-kit-dev": "4.0.0-extensibility-preview.3",
-    "@salesforce/pwa-kit-extension-sdk": "4.0.0-extensibility-preview.3",
-    "@salesforce/pwa-kit-react-sdk": "4.0.0-extensibility-preview.3",
-    "@salesforce/pwa-kit-runtime": "4.0.0-extensibility-preview.3",
-    "@tanstack/react-query": "^4.28.0",
-    "@types/loadable__component": "~5.13.4",
-    "@types/react": "~18.2.0",
-    "@types/react-dom": "~18.2.1",
-    "@types/react-router-dom": "~5.3.3",
-    "framer-motion": "^11.5.4",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-helmet": "^6.1.0",
-    "react-intl": "^5.25.1",
-    "react-router-dom": "^5.3.4",
-    "typescript": "4.9.5"
-  },
-  "engines": {
-    "node": "^16.11.0 || ^18.0.0 || ^20.0.0",
-    "npm": "^8.0.0 || ^9.0.0 || ^10.0.0"
-  },
-  "mobify": {
-    "app": {
-      "extensions": [
-        [
-          "@salesforce/extension-chakra-storefront"
-        ]
-      ]
+    "name": "typescript-minimal",
+    "version": "4.0.0-extensibility-preview.3",
+    "private": true,
+    "scripts": {
+        "analyze-build": "cross-env MOBIFY_ANALYZE=true npm run build",
+        "build": "pwa-kit-dev build",
+        "format": "pwa-kit-dev format \"**/*.{js,jsx,ts,tsx}\"",
+        "lint": "npm run lint:js && npm run typecheck",
+        "lint:fix": "npm run lint:js -- --fix",
+        "lint:js": "pwa-kit-dev lint \"**/*.{js,jsx,ts,tsx}\"",
+        "push": "npm run build && pwa-kit-dev push",
+        "save-credentials": "pwa-kit-dev save-credentials",
+        "start": "pwa-kit-dev start",
+        "start:inspect": "npm run start -- --inspect",
+        "tail-logs": "pwa-kit-dev tail-logs",
+        "test": "pwa-kit-dev test",
+        "typecheck": "tsc --noEmit"
     },
-    "ssrEnabled": true,
-    "ssrOnly": [
-      "ssr.js",
-      "ssr.js.map",
-      "node_modules/**/*.*"
-    ],
-    "ssrShared": [
-      "static/favicon.ico",
-      "static/robots.txt",
-      "**/*.js",
-      "**/*.js.map",
-      "**/*.json"
-    ],
-    "ssrParameters": {
-      "ssrFunctionNodeVersion": "20.x",
-      "proxyConfigs": [
-        {
-          "host": "kv7kzm78.api.commercecloud.salesforce.com",
-          "path": "api"
+    "devDependencies": {
+        "@chakra-ui/react": "^2.8.2",
+        "@emotion/react": "^11.13.3",
+        "@emotion/styled": "^11.13.0",
+        "@loadable/component": "^5.15.3",
+        "@salesforce/extension-chakra-storefront": "0.1.0-extensibility-preview.3",
+        "@salesforce/pwa-kit-dev": "4.0.0-extensibility-preview.3",
+        "@salesforce/pwa-kit-extension-sdk": "4.0.0-extensibility-preview.3",
+        "@salesforce/pwa-kit-react-sdk": "4.0.0-extensibility-preview.3",
+        "@salesforce/pwa-kit-runtime": "4.0.0-extensibility-preview.3",
+        "@tanstack/react-query": "^4.28.0",
+        "@types/loadable__component": "~5.13.4",
+        "@types/react": "~18.2.0",
+        "@types/react-dom": "~18.2.1",
+        "@types/react-router-dom": "~5.3.3",
+        "framer-motion": "^11.5.4",
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "react-helmet": "^6.1.0",
+        "react-intl": "^5.25.1",
+        "react-router-dom": "^5.3.4",
+        "typescript": "4.9.5"
+    },
+    "engines": {
+        "node": "^16.11.0 || ^18.0.0 || ^20.0.0",
+        "npm": "^8.0.0 || ^9.0.0 || ^10.0.0"
+    },
+    "mobify": {
+        "app": {
+            "extensions": [
+                [
+                    "@salesforce/extension-chakra-storefront"
+                ]
+            ]
         },
-        {
-          "host": "zzrf-001.dx.commercecloud.salesforce.com",
-          "path": "ocapi"
+        "ssrEnabled": true,
+        "ssrOnly": [
+            "ssr.js",
+            "ssr.js.map",
+            "node_modules/**/*.*"
+        ],
+        "ssrShared": [
+            "static/favicon.ico",
+            "static/robots.txt",
+            "**/*.js",
+            "**/*.js.map",
+            "**/*.json"
+        ],
+        "ssrParameters": {
+            "ssrFunctionNodeVersion": "20.x",
+            "proxyConfigs": [
+                {
+                    "host": "kv7kzm78.api.commercecloud.salesforce.com",
+                    "path": "api"
+                },
+                {
+                    "host": "zzrf-001.dx.commercecloud.salesforce.com",
+                    "path": "ocapi"
+                }
+            ]
         }
-      ]
     }
-  }
 }
